Remove all support for 9.1 and lower.
It will no longer work with the new semverrange type, anyway, so just
remove the Makefile support for it. Instead, check for 9.2 or higher
and raise an error if it's lower. This lets the Makefile be quite a
bit simpler, and we no longer need to do filtering tricks to include
or exclude files -- we just remove the uninstall script and call it
good. Resolves #31.