Revamp send/receive and test it.
Follow ltree's example by prepending a version and using the pg_msg functions.
Add tests. A little hinky to use `\copy` to write to a file and then read it
back in, but it is the only way I could figure out to test the receive
function, which binary COPY FROM uses.

Increment to v0.32.0, update the copyright date, and test on Postgres 15.