Use domain-qualified name for extension
An experiment to see whether the current PostgreSQL extension
installation and configuration can be abused to avoid namespace
collision between extensions with the same names. This shows that
for relocatable extensions, at least, it can be done, but it's
pretty ugly: Everything has to refer to the full extension name ---
here `github.com+theory+pg-semver` --- including:

*   The Control file
*   SQL files
*   The `MODULERDIR` variable in the `Makefile`, to install files into a
    directory with the long extension name
*   The `directory` entry in the control file to match `MODULERDIR`

This change doesn't actually rename all of the `sql/*.sql` files, but
updates the `Makefile` to generate the
`sql/github.com+theory+pg-semver--$VERSION.sql` file

It's possible a lot of this could be automated in the `Makefile`:
replace a short name with a longer name, move and rename files, etc.
But it'd be kind of a pain. The alternative is to rename everything
appropriately in the repository, but that'd be ugly. Besides, changing
the extension name like this means that older versions would never be
properly upgraded.

It'd be nicer if PostgreSQL itself were updated to support some sort of
schema packaging, where everything could go into one directory named for
the repository. That's probably the smarter long-term goal.