Use MODULE_PATHNAME and CREATE EXTENSION
Use `MODULE_PATHNAME` in the SQL files to refer to the static library to
build. This allows for the library to be renamed --- e.g., by a
packaging system --- without having to change the source code, just the
control file.

Also use `CREATE EXTENSION` in the tests instead of loading the
`sql/semver.sql` file directly. The latter was required before extension
support was added to Postgres in 9.1, and 9.2 has been required since
v0.20.0 back in 2018.

Also note the change to the module directory in the previous commit
(88b3abd), ignore generated `*.bin` files, and update the copyright
year.