POC Trunk binary distribution format
`trunk.mk` copies a bunch of the `install` target from PGXS and modifies
it to install into a organized according to the [proposed format]. The
adds the `make` variables `DISTVERSION`, `LICENSE`, and `LANGUAGE`, but
otherwise depends on variables defined by PGXS. It uses `jq` to build
`trunk.json`, `shasum` to build the `digests` file, and `tar` to create
the `*.trunk` artifact as a tarball.

`install_trunk` is a simple shell script to demonstrate unpacking,
validating, and installing a trunk file. It too relies on `jq`,
`shasum`, and `tar`, and uses `rsync` to do the actual installation. It
also does basic trunk version, platform, and Postgres version
validation.

  [proposed format]: https://github.com/orgs/pgxn/discussions/2